requirements.txt
aiogram==3.15.0
python-dotenv==1.0.1
python-docx==1.1.2
aiosqlite==0.20.0
email-validator==2.2.0